Fix conversion error

+++ src/object/player.cpp
@@ -2759,7 +2759,7 @@ Player::stop_rolling(bool violent)
   {
     for (int i = 0; i < 5; i++)
     {
-      Vector pspeed = Vector(graphicsRandom.randf(-100.f, 100.f)*(i-2), graphicsRandom.randf(-200.f, -150.f));
+      Vector pspeed = Vector(graphicsRandom.randf(-100.f, 100.f)*(static_cast<float>(i)-2), graphicsRandom.randf(-200.f, -150.f));
       Vector paccel = Vector(0, 1000.f + graphicsRandom.randf(-100.f, 100.f));
       Sector::get().add<SpriteParticle>(
         "images/particles/rock.sprite", "rock-"+std::to_string(i),
